+++ tests/otel_integrations/test_fastapi.py
@@ -32,7 +32,7 @@ def test_missing_opentelemetry_dependency() -> None:
         with pytest.raises(RuntimeError) as exc_info:
             importlib.reload(logfire._internal.integrations.fastapi)
         assert str(exc_info.value) == snapshot("""\
-The `logfire.instrument_fastapi()` requires the `opentelemetry-instrumentation-fastapi` package.
+The `logfire.instrument_fastapi()` method requires the `opentelemetry-instrumentation-fastapi` package.
 You can install this with:
     pip install 'logfire[fastapi]'\
 """)
@@ -186,8 +186,8 @@ def test_bad_dependency_route(client: TestClient, exporter: TestExporter) -> Non
                 'name': 'GET /bad_dependency_route/{good} http send response.start',
                 'context': {'trace_id': 1, 'span_id': 5, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
-                'start_time': 8000000000,
-                'end_time': 9000000000,
+                'start_time': 9000000000,
+                'end_time': 10000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'GET /bad_dependency_route/{good} http send response.start',
@@ -202,8 +202,8 @@ def test_bad_dependency_route(client: TestClient, exporter: TestExporter) -> Non
                 'name': 'GET /bad_dependency_route/{good} http send response.body',
                 'context': {'trace_id': 1, 'span_id': 7, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
-                'start_time': 10000000000,
-                'end_time': 11000000000,
+                'start_time': 11000000000,
+                'end_time': 12000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'GET /bad_dependency_route/{good} http send response.body',
@@ -216,7 +216,7 @@ def test_bad_dependency_route(client: TestClient, exporter: TestExporter) -> Non
                 'context': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
                 'parent': None,
                 'start_time': 1000000000,
-                'end_time': 12000000000,
+                'end_time': 13000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'GET /bad_dependency_route/good_value',
@@ -257,7 +257,7 @@ def test_bad_dependency_route(client: TestClient, exporter: TestExporter) -> Non
                 'events': [
                     {
                         'name': 'exception',
-                        'timestamp': 5000000000,
+                        'timestamp': 8000000000,
                         'attributes': {
                             'exception.type': 'ValueError',
                             'exception.message': 'bad dependency',
@@ -1491,8 +1491,8 @@ def test_fastapi_unhandled_exception(client: TestClient, exporter: TestExporter)
                 'name': 'GET /exception http send response.start',
                 'context': {'trace_id': 1, 'span_id': 7, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
-                'start_time': 12000000000,
-                'end_time': 13000000000,
+                'start_time': 13000000000,
+                'end_time': 14000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'GET /exception http send response.start',
@@ -1507,8 +1507,8 @@ def test_fastapi_unhandled_exception(client: TestClient, exporter: TestExporter)
                 'name': 'GET /exception http send response.body',
                 'context': {'trace_id': 1, 'span_id': 9, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
-                'start_time': 14000000000,
-                'end_time': 15000000000,
+                'start_time': 15000000000,
+                'end_time': 16000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'GET /exception http send response.body',
@@ -1521,7 +1521,7 @@ def test_fastapi_unhandled_exception(client: TestClient, exporter: TestExporter)
                 'context': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
                 'parent': None,
                 'start_time': 1000000000,
-                'end_time': 16000000000,
+                'end_time': 17000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'GET /exception',
@@ -1561,7 +1561,7 @@ def test_fastapi_unhandled_exception(client: TestClient, exporter: TestExporter)
                 'events': [
                     {
                         'name': 'exception',
-                        'timestamp': 9000000000,
+                        'timestamp': 12000000000,
                         'attributes': {
                             'exception.type': 'ValueError',
                             'exception.message': 'test exception',
@@ -1709,6 +1709,7 @@ def test_fastapi_handled_exception(client: TestClient, exporter: TestExporter) -
                             'exception.message': '[]',
                             'exception.stacktrace': 'fastapi.exceptions.RequestValidationError: []',
                             'exception.escaped': 'False',
+                            'recorded_by_logfire_fastapi': True,
                         },
                     }
                 ],
@@ -1892,6 +1893,7 @@ def make_request_hook_spans(record_send_receive: bool):
     def server_request_hook(span: Any, _scope: dict[str, Any]):
         logfire.info('server_request_hook')
         set_user_attributes_on_raw_span(span, {'attr_key': 'attr_val'})
+        span.add_event('server_request_hook')
 
     with logfire.instrument_fastapi(
         app,
@@ -1930,8 +1932,8 @@ def test_request_hooks_without_send_receiev_spans(exporter: TestExporter):
                 'name': 'client_request_hook',
                 'context': {'trace_id': 1, 'span_id': 4, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
-                'start_time': 6000000000,
-                'end_time': 6000000000,
+                'start_time': 7000000000,
+                'end_time': 7000000000,
                 'attributes': {
                     'logfire.span_type': 'log',
                     'logfire.level_num': 9,
@@ -1946,8 +1948,8 @@ def test_request_hooks_without_send_receiev_spans(exporter: TestExporter):
                 'name': 'client_response_hook',
                 'context': {'trace_id': 1, 'span_id': 5, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
-                'start_time': 8000000000,
-                'end_time': 8000000000,
+                'start_time': 9000000000,
+                'end_time': 9000000000,
                 'attributes': {
                     'logfire.span_type': 'log',
                     'logfire.level_num': 9,
@@ -1962,8 +1964,8 @@ def test_request_hooks_without_send_receiev_spans(exporter: TestExporter):
                 'name': 'client_response_hook',
                 'context': {'trace_id': 1, 'span_id': 6, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
-                'start_time': 9000000000,
-                'end_time': 9000000000,
+                'start_time': 10000000000,
+                'end_time': 10000000000,
                 'attributes': {
                     'logfire.span_type': 'log',
                     'logfire.level_num': 9,
@@ -1979,7 +1981,7 @@ def test_request_hooks_without_send_receiev_spans(exporter: TestExporter):
                 'context': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
                 'parent': None,
                 'start_time': 1000000000,
-                'end_time': 10000000000,
+                'end_time': 11000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'POST /echo_body',
@@ -2007,16 +2009,17 @@ def test_request_hooks_without_send_receiev_spans(exporter: TestExporter):
                     'attr_key': 'attr_val',
                     'fastapi.route.name': 'echo_body',
                     'fastapi.route.operation_id': 'null',
-                    'fastapi.arguments.start_timestamp': '1970-01-01T00:00:03.000000Z',
-                    'fastapi.arguments.end_timestamp': '1970-01-01T00:00:04.000000Z',
+                    'fastapi.arguments.start_timestamp': '1970-01-01T00:00:04.000000Z',
+                    'fastapi.arguments.end_timestamp': '1970-01-01T00:00:05.000000Z',
                     'fastapi.arguments.values': '{}',
                     'fastapi.arguments.errors': '[]',
-                    'fastapi.endpoint_function.start_timestamp': '1970-01-01T00:00:05.000000Z',
-                    'fastapi.endpoint_function.end_timestamp': '1970-01-01T00:00:07.000000Z',
+                    'fastapi.endpoint_function.start_timestamp': '1970-01-01T00:00:06.000000Z',
+                    'fastapi.endpoint_function.end_timestamp': '1970-01-01T00:00:08.000000Z',
                     'logfire.json_schema': '{"type":"object","properties":{"attr_key":{},"fastapi.route.name":{},"fastapi.route.operation_id":{"type":"null"},"fastapi.arguments.values":{"type":"object"},"fastapi.arguments.errors":{"type":"array"}}}',
                     'http.status_code': 200,
                     'http.response.status_code': 200,
                 },
+                'events': [{'name': 'server_request_hook', 'timestamp': 3000000000}],
             },
         ]
     )
@@ -2046,8 +2049,8 @@ def test_request_hooks_with_send_receive_spans(exporter: TestExporter):
                 'name': 'client_request_hook',
                 'context': {'trace_id': 1, 'span_id': 6, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 4, 'is_remote': False},
-                'start_time': 7000000000,
-                'end_time': 7000000000,
+                'start_time': 8000000000,
+                'end_time': 8000000000,
                 'attributes': {
                     'logfire.span_type': 'log',
                     'logfire.level_num': 9,
@@ -2062,8 +2065,8 @@ def test_request_hooks_with_send_receive_spans(exporter: TestExporter):
                 'name': 'POST /echo_body http receive request',
                 'context': {'trace_id': 1, 'span_id': 4, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
-                'start_time': 6000000000,
-                'end_time': 8000000000,
+                'start_time': 7000000000,
+                'end_time': 9000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'POST /echo_body http receive request',
@@ -2075,8 +2078,8 @@ def test_request_hooks_with_send_receive_spans(exporter: TestExporter):
                 'name': 'client_response_hook',
                 'context': {'trace_id': 1, 'span_id': 9, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 7, 'is_remote': False},
-                'start_time': 11000000000,
-                'end_time': 11000000000,
+                'start_time': 12000000000,
+                'end_time': 12000000000,
                 'attributes': {
                     'logfire.span_type': 'log',
                     'logfire.level_num': 9,
@@ -2091,8 +2094,8 @@ def test_request_hooks_with_send_receive_spans(exporter: TestExporter):
                 'name': 'POST /echo_body http send response.start',
                 'context': {'trace_id': 1, 'span_id': 7, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
-                'start_time': 10000000000,
-                'end_time': 12000000000,
+                'start_time': 11000000000,
+                'end_time': 13000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'POST /echo_body http send response.start',
@@ -2106,8 +2109,8 @@ def test_request_hooks_with_send_receive_spans(exporter: TestExporter):
                 'name': 'client_response_hook',
                 'context': {'trace_id': 1, 'span_id': 12, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 10, 'is_remote': False},
-                'start_time': 14000000000,
-                'end_time': 14000000000,
+                'start_time': 15000000000,
+                'end_time': 15000000000,
                 'attributes': {
                     'logfire.span_type': 'log',
                     'logfire.level_num': 9,
@@ -2122,8 +2125,8 @@ def test_request_hooks_with_send_receive_spans(exporter: TestExporter):
                 'name': 'POST /echo_body http send response.body',
                 'context': {'trace_id': 1, 'span_id': 10, 'is_remote': False},
                 'parent': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
-                'start_time': 13000000000,
-                'end_time': 15000000000,
+                'start_time': 14000000000,
+                'end_time': 16000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'POST /echo_body http send response.body',
@@ -2136,7 +2139,7 @@ def test_request_hooks_with_send_receive_spans(exporter: TestExporter):
                 'context': {'trace_id': 1, 'span_id': 1, 'is_remote': False},
                 'parent': None,
                 'start_time': 1000000000,
-                'end_time': 16000000000,
+                'end_time': 17000000000,
                 'attributes': {
                     'logfire.span_type': 'span',
                     'logfire.msg': 'POST /echo_body',
@@ -2164,16 +2167,22 @@ def test_request_hooks_with_send_receive_spans(exporter: TestExporter):
                     'attr_key': 'attr_val',
                     'fastapi.route.name': 'echo_body',
                     'fastapi.route.operation_id': 'null',
-                    'fastapi.arguments.start_timestamp': '1970-01-01T00:00:03.000000Z',
-                    'fastapi.arguments.end_timestamp': '1970-01-01T00:00:04.000000Z',
+                    'fastapi.arguments.start_timestamp': '1970-01-01T00:00:04.000000Z',
+                    'fastapi.arguments.end_timestamp': '1970-01-01T00:00:05.000000Z',
                     'fastapi.arguments.values': '{}',
                     'fastapi.arguments.errors': '[]',
-                    'fastapi.endpoint_function.start_timestamp': '1970-01-01T00:00:05.000000Z',
-                    'fastapi.endpoint_function.end_timestamp': '1970-01-01T00:00:09.000000Z',
+                    'fastapi.endpoint_function.start_timestamp': '1970-01-01T00:00:06.000000Z',
+                    'fastapi.endpoint_function.end_timestamp': '1970-01-01T00:00:10.000000Z',
                     'logfire.json_schema': '{"type":"object","properties":{"attr_key":{},"fastapi.route.name":{},"fastapi.route.operation_id":{"type":"null"},"fastapi.arguments.values":{"type":"object"},"fastapi.arguments.errors":{"type":"array"}}}',
                     'http.status_code': 200,
                     'http.response.status_code': 200,
                 },
+                'events': [
+                    {
+                        'name': 'server_request_hook',
+                        'timestamp': 3000000000,
+                    }
+                ],
             },
         ]
     )